+++ fabric/src/main/java/com/lx862/mtrscripting/core/ParsedScript.java
@@ -42,7 +42,8 @@ public class ParsedScript {
             scope.put("StateTracker", scope, new NativeJavaClass(scope, StateTracker.class));
             scope.put("CycleTracker", scope, new NativeJavaClass(scope, CycleTracker.class));
             scope.put("RateLimit", scope, new NativeJavaClass(scope, RateLimit.class));
-            scope.put("HttpUtil", scope, new NativeJavaClass(scope, HttpUtil.class));
+            scope.put("Networking", scope, new NativeJavaClass(scope, NetworkingUtil.class));
+            scope.put("Files", scope, new NativeJavaClass(scope, FilesUtil.class));
 
             scope.put("Matrices", scope, new NativeJavaClass(scope, Matrices.class));